Update OBS sdk to v3.21.12
