+++ openstack/obs/auth.go
@@ -285,44 +285,6 @@ func attachHeaders(headers map[string][]string, isObs bool) string {
 	return strings.Join(stringToSign, "\n")
 }
 
-func getV2StringToSign(method, canonicalizedURL string, headers map[string][]string, isObs bool) string {
-	stringToSign := strings.Join([]string{method, "\n", attachHeaders(headers, isObs), "\n", canonicalizedURL}, "")
-
-	var isSecurityToken bool
-	var securityToken []string
-	if isObs {
-		securityToken, isSecurityToken = headers[HEADER_STS_TOKEN_OBS]
-	} else {
-		securityToken, isSecurityToken = headers[HEADER_STS_TOKEN_AMZ]
-	}
-	var query []string
-	if !isSecurityToken {
-		parmas := strings.Split(canonicalizedURL, "?")
-		if len(parmas) > 1 {
-			query = strings.Split(parmas[1], "&")
-			for _, value := range query {
-				if strings.HasPrefix(value, HEADER_STS_TOKEN_AMZ+"=") || strings.HasPrefix(value, HEADER_STS_TOKEN_OBS+"=") {
-					if value[len(HEADER_STS_TOKEN_AMZ)+1:] != "" {
-						securityToken = []string{value[len(HEADER_STS_TOKEN_AMZ)+1:]}
-						isSecurityToken = true
-					}
-				}
-			}
-		}
-	}
-	logStringToSign := stringToSign
-	if isSecurityToken && len(securityToken) > 0 {
-		logStringToSign = strings.Replace(logStringToSign, securityToken[0], "******", -1)
-	}
-	doLog(LEVEL_DEBUG, "The v2 auth stringToSign:\n%s", logStringToSign)
-	return stringToSign
-}
-
-func v2Auth(ak, sk, method, canonicalizedURL string, headers map[string][]string, isObs bool) map[string]string {
-	stringToSign := getV2StringToSign(method, canonicalizedURL, headers, isObs)
-	return map[string]string{"Signature": Base64Encode(HmacSha1([]byte(sk), []byte(stringToSign)))}
-}
-
 func getScope(region, shortDate string) string {
 	return fmt.Sprintf("%s/%s/%s/%s", shortDate, region, V4_SERVICE_NAME, V4_SERVICE_SUFFIX)
 }
@@ -330,69 +292,6 @@ func getScope(region, shortDate string) string {
 func getCredential(ak, region, shortDate string) (string, string) {
 	scope := getScope(region, shortDate)
 	return fmt.Sprintf("%s/%s", ak, scope), scope
-}
-
-func getV4StringToSign(method, canonicalizedURL, queryURL, scope, longDate, payload string, signedHeaders []string, headers map[string][]string) string {
-	canonicalRequest := make([]string, 0, 10+len(signedHeaders)*4)
-	canonicalRequest = append(canonicalRequest, method)
-	canonicalRequest = append(canonicalRequest, "\n")
-	canonicalRequest = append(canonicalRequest, canonicalizedURL)
-	canonicalRequest = append(canonicalRequest, "\n")
-	canonicalRequest = append(canonicalRequest, queryURL)
-	canonicalRequest = append(canonicalRequest, "\n")
-
-	for _, signedHeader := range signedHeaders {
-		values, _ := headers[signedHeader]
-		for _, value := range values {
-			canonicalRequest = append(canonicalRequest, signedHeader)
-			canonicalRequest = append(canonicalRequest, ":")
-			canonicalRequest = append(canonicalRequest, value)
-			canonicalRequest = append(canonicalRequest, "\n")
-		}
-	}
-	canonicalRequest = append(canonicalRequest, "\n")
-	canonicalRequest = append(canonicalRequest, strings.Join(signedHeaders, ";"))
-	canonicalRequest = append(canonicalRequest, "\n")
-	canonicalRequest = append(canonicalRequest, payload)
-
-	_canonicalRequest := strings.Join(canonicalRequest, "")
-
-	var isSecurityToken bool
-	var securityToken []string
-	if securityToken, isSecurityToken = headers[HEADER_STS_TOKEN_OBS]; !isSecurityToken {
-		securityToken, isSecurityToken = headers[HEADER_STS_TOKEN_AMZ]
-	}
-	var query []string
-	if !isSecurityToken {
-		query = strings.Split(queryURL, "&")
-		for _, value := range query {
-			if strings.HasPrefix(value, HEADER_STS_TOKEN_AMZ+"=") || strings.HasPrefix(value, HEADER_STS_TOKEN_OBS+"=") {
-				if value[len(HEADER_STS_TOKEN_AMZ)+1:] != "" {
-					securityToken = []string{value[len(HEADER_STS_TOKEN_AMZ)+1:]}
-					isSecurityToken = true
-				}
-			}
-		}
-	}
-	logCanonicalRequest := _canonicalRequest
-	if isSecurityToken && len(securityToken) > 0 {
-		logCanonicalRequest = strings.Replace(logCanonicalRequest, securityToken[0], "******", -1)
-	}
-	doLog(LEVEL_DEBUG, "The v4 auth canonicalRequest:\n%s", logCanonicalRequest)
-
-	stringToSign := make([]string, 0, 7)
-	stringToSign = append(stringToSign, V4_HASH_PREFIX)
-	stringToSign = append(stringToSign, "\n")
-	stringToSign = append(stringToSign, longDate)
-	stringToSign = append(stringToSign, "\n")
-	stringToSign = append(stringToSign, scope)
-	stringToSign = append(stringToSign, "\n")
-	stringToSign = append(stringToSign, HexSha256([]byte(_canonicalRequest)))
-
-	_stringToSign := strings.Join(stringToSign, "")
-
-	doLog(LEVEL_DEBUG, "The v4 auth stringToSign:\n%s", _stringToSign)
-	return _stringToSign
 }
 
 func getSignedHeaders(headers map[string][]string) ([]string, map[string][]string) {
@@ -418,60 +317,4 @@ func getSignature(stringToSign, sk, region, shortDate string) string {
 	key = HmacSha256(key, []byte(V4_SERVICE_NAME))
 	key = HmacSha256(key, []byte(V4_SERVICE_SUFFIX))
 	return Hex(HmacSha256(key, []byte(stringToSign)))
-}
-
-// V4Auth is a wrapper for v4Auth
-func V4Auth(ak, sk, region, method, canonicalizedURL, queryURL string, headers map[string][]string) map[string]string {
-	return v4Auth(ak, sk, region, method, canonicalizedURL, queryURL, headers)
-}
-
-func v4Auth(ak, sk, region, method, canonicalizedURL, queryURL string, headers map[string][]string) map[string]string {
-	var t time.Time
-	if val, ok := headers[HEADER_DATE_AMZ]; ok {
-		var err error
-		t, err = time.Parse(LONG_DATE_FORMAT, val[0])
-		if err != nil {
-			t = time.Now().UTC()
-		}
-	} else if val, ok := headers[PARAM_DATE_AMZ_CAMEL]; ok {
-		var err error
-		t, err = time.Parse(LONG_DATE_FORMAT, val[0])
-		if err != nil {
-			t = time.Now().UTC()
-		}
-	} else if val, ok := headers[HEADER_DATE_CAMEL]; ok {
-		var err error
-		t, err = time.Parse(RFC1123_FORMAT, val[0])
-		if err != nil {
-			t = time.Now().UTC()
-		}
-	} else if val, ok := headers[strings.ToLower(HEADER_DATE_CAMEL)]; ok {
-		var err error
-		t, err = time.Parse(RFC1123_FORMAT, val[0])
-		if err != nil {
-			t = time.Now().UTC()
-		}
-	} else {
-		t = time.Now().UTC()
-	}
-	shortDate := t.Format(SHORT_DATE_FORMAT)
-	longDate := t.Format(LONG_DATE_FORMAT)
-
-	signedHeaders, _headers := getSignedHeaders(headers)
-
-	credential, scope := getCredential(ak, region, shortDate)
-
-	payload := UNSIGNED_PAYLOAD
-	if val, ok := headers[HEADER_CONTENT_SHA256_AMZ]; ok {
-		payload = val[0]
-	}
-	stringToSign := getV4StringToSign(method, canonicalizedURL, queryURL, scope, longDate, payload, signedHeaders, _headers)
-
-	signature := getSignature(stringToSign, sk, region, shortDate)
-
-	ret := make(map[string]string, 3)
-	ret["Credential"] = credential
-	ret["SignedHeaders"] = strings.Join(signedHeaders, ";")
-	ret["Signature"] = signature
-	return ret
 }
