+++ openstack/obs/client_bucket.go
@@ -15,86 +15,8 @@ package obs
 import (
 	"errors"
 	"fmt"
-	"io"
-	"net/http"
-	"os"
-	"sort"
 	"strings"
 )
-
-// ObsClient defines OBS client.
-type ObsClient struct {
-	conf       *config
-	httpClient *http.Client
-}
-
-// New creates a new ObsClient instance.
-func New(ak, sk, endpoint string, configurers ...configurer) (*ObsClient, error) {
-	conf := &config{endpoint: endpoint}
-	conf.securityProviders = make([]securityProvider, 0, 3)
-	conf.securityProviders = append(conf.securityProviders, NewBasicSecurityProvider(ak, sk, ""))
-
-	conf.maxRetryCount = -1
-	conf.maxRedirectCount = -1
-	for _, configurer := range configurers {
-		configurer(conf)
-	}
-
-	if err := conf.initConfigWithDefault(); err != nil {
-		return nil, err
-	}
-	err := conf.getTransport()
-	if err != nil {
-		return nil, err
-	}
-
-	if isWarnLogEnabled() {
-		info := make([]string, 3)
-		info[0] = fmt.Sprintf("[OBS SDK Version=%s", obsSdkVersion)
-		info[1] = fmt.Sprintf("Endpoint=%s", conf.endpoint)
-		accessMode := "Virtual Hosting"
-		if conf.pathStyle {
-			accessMode = "Path"
-		}
-		info[2] = fmt.Sprintf("Access Mode=%s]", accessMode)
-		doLog(LEVEL_WARN, strings.Join(info, "];["))
-	}
-	doLog(LEVEL_DEBUG, "Create obsclient with config:\n%s\n", conf)
-	obsClient := &ObsClient{conf: conf, httpClient: &http.Client{Transport: conf.transport, CheckRedirect: checkRedirectFunc}}
-	return obsClient, nil
-}
-
-// Refresh refreshes ak, sk and securityToken for obsClient.
-func (obsClient ObsClient) Refresh(ak, sk, securityToken string) {
-	for _, sp := range obsClient.conf.securityProviders {
-		if bsp, ok := sp.(*BasicSecurityProvider); ok {
-			bsp.refresh(strings.TrimSpace(ak), strings.TrimSpace(sk), strings.TrimSpace(securityToken))
-			break
-		}
-	}
-}
-
-func (obsClient ObsClient) getSecurity() securityHolder {
-	if obsClient.conf.securityProviders != nil {
-		for _, sp := range obsClient.conf.securityProviders {
-			if sp == nil {
-				continue
-			}
-			sh := sp.getSecurity()
-			if sh.ak != "" && sh.sk != "" {
-				return sh
-			}
-		}
-	}
-	return emptySecurityHolder
-}
-
-// Close closes ObsClient.
-func (obsClient ObsClient) Close() {
-	obsClient.httpClient = nil
-	obsClient.conf.transport.CloseIdleConnections()
-	obsClient.conf = nil
-}
 
 // ListBuckets lists buckets.
 //
@@ -191,79 +113,6 @@ func (obsClient ObsClient) GetBucketStoragePolicy(bucketName string, extensions 
 	return obsClient.getBucketStoragePolicyS3(bucketName, extensions)
 }
 
-// ListObjects lists objects in a bucket.
-//
-// You can use this API to list objects in a bucket. By default, a maximum of 1000 objects are listed.
-func (obsClient ObsClient) ListObjects(input *ListObjectsInput, extensions ...extensionOptions) (output *ListObjectsOutput, err error) {
-	if input == nil {
-		return nil, errors.New("ListObjectsInput is nil")
-	}
-	output = &ListObjectsOutput{}
-	err = obsClient.doActionWithBucket("ListObjects", HTTP_GET, input.Bucket, input, output, extensions)
-	if err != nil {
-		output = nil
-	} else {
-		if location, ok := output.ResponseHeaders[HEADER_BUCKET_REGION]; ok {
-			output.Location = location[0]
-		}
-		if output.EncodingType == "url" {
-			err = decodeListObjectsOutput(output)
-			if err != nil {
-				doLog(LEVEL_ERROR, "Failed to get ListObjectsOutput with error: %v.", err)
-				output = nil
-			}
-		}
-	}
-	return
-}
-
-// ListVersions lists versioning objects in a bucket.
-//
-// You can use this API to list versioning objects in a bucket. By default, a maximum of 1000 versioning objects are listed.
-func (obsClient ObsClient) ListVersions(input *ListVersionsInput, extensions ...extensionOptions) (output *ListVersionsOutput, err error) {
-	if input == nil {
-		return nil, errors.New("ListVersionsInput is nil")
-	}
-	output = &ListVersionsOutput{}
-	err = obsClient.doActionWithBucket("ListVersions", HTTP_GET, input.Bucket, input, output, extensions)
-	if err != nil {
-		output = nil
-	} else {
-		if location, ok := output.ResponseHeaders[HEADER_BUCKET_REGION]; ok {
-			output.Location = location[0]
-		}
-		if output.EncodingType == "url" {
-			err = decodeListVersionsOutput(output)
-			if err != nil {
-				doLog(LEVEL_ERROR, "Failed to get ListVersionsOutput with error: %v.", err)
-				output = nil
-			}
-		}
-	}
-	return
-}
-
-// ListMultipartUploads lists the multipart uploads.
-//
-// You can use this API to list the multipart uploads that are initialized but not combined or aborted in a specified bucket.
-func (obsClient ObsClient) ListMultipartUploads(input *ListMultipartUploadsInput, extensions ...extensionOptions) (output *ListMultipartUploadsOutput, err error) {
-	if input == nil {
-		return nil, errors.New("ListMultipartUploadsInput is nil")
-	}
-	output = &ListMultipartUploadsOutput{}
-	err = obsClient.doActionWithBucket("ListMultipartUploads", HTTP_GET, input.Bucket, input, output, extensions)
-	if err != nil {
-		output = nil
-	} else if output.EncodingType == "url" {
-		err = decodeListMultipartUploadsOutput(output)
-		if err != nil {
-			doLog(LEVEL_ERROR, "Failed to get ListMultipartUploadsOutput with error: %v.", err)
-			output = nil
-		}
-	}
-	return
-}
-
 // SetBucketQuota sets the bucket quota.
 //
 // You can use this API to set the bucket quota. A bucket quota must be expressed in bytes and the maximum value is 2^63-1.
@@ -303,21 +152,6 @@ func (obsClient ObsClient) HeadBucket(bucketName string, extensions ...extension
 	return
 }
 
-// HeadObject checks whether an object exists.
-//
-// You can use this API to check whether an object exists.
-func (obsClient ObsClient) HeadObject(input *HeadObjectInput, extensions ...extensionOptions) (output *BaseModel, err error) {
-	if input == nil {
-		return nil, errors.New("HeadObjectInput is nil")
-	}
-	output = &BaseModel{}
-	err = obsClient.doActionWithBucketAndKey("HeadObject", HTTP_HEAD, input.Bucket, input.Key, input, output, extensions)
-	if err != nil {
-		output = nil
-	}
-	return
-}
-
 // GetBucketMetadata gets the metadata of a bucket.
 //
 // You can use this API to send a HEAD request to a bucket to obtain the bucket
@@ -329,18 +163,6 @@ func (obsClient ObsClient) GetBucketMetadata(input *GetBucketMetadataInput, exte
 		output = nil
 	} else {
 		ParseGetBucketMetadataOutput(output)
-	}
-	return
-}
-
-// SetObjectMetadata sets object metadata.
-func (obsClient ObsClient) SetObjectMetadata(input *SetObjectMetadataInput, extensions ...extensionOptions) (output *SetObjectMetadataOutput, err error) {
-	output = &SetObjectMetadataOutput{}
-	err = obsClient.doActionWithBucketAndKey("SetObjectMetadata", HTTP_PUT, input.Bucket, input.Key, input, output, extensions)
-	if err != nil {
-		output = nil
-	} else {
-		ParseSetObjectMetadataOutput(output)
 	}
 	return
 }
@@ -798,464 +620,6 @@ func (obsClient ObsClient) getBucketNotificationS3(bucketName string, extensions
 	return
 }
 
-// DeleteObject deletes an object.
-//
-// You can use this API to delete an object from a specified bucket.
-func (obsClient ObsClient) DeleteObject(input *DeleteObjectInput, extensions ...extensionOptions) (output *DeleteObjectOutput, err error) {
-	if input == nil {
-		return nil, errors.New("DeleteObjectInput is nil")
-	}
-	output = &DeleteObjectOutput{}
-	err = obsClient.doActionWithBucketAndKey("DeleteObject", HTTP_DELETE, input.Bucket, input.Key, input, output, extensions)
-	if err != nil {
-		output = nil
-	} else {
-		ParseDeleteObjectOutput(output)
-	}
-	return
-}
-
-// DeleteObjects deletes objects in a batch.
-//
-// You can use this API to batch delete objects from a specified bucket.
-func (obsClient ObsClient) DeleteObjects(input *DeleteObjectsInput, extensions ...extensionOptions) (output *DeleteObjectsOutput, err error) {
-	if input == nil {
-		return nil, errors.New("DeleteObjectsInput is nil")
-	}
-	output = &DeleteObjectsOutput{}
-	err = obsClient.doActionWithBucket("DeleteObjects", HTTP_POST, input.Bucket, input, output, extensions)
-	if err != nil {
-		output = nil
-	} else if output.EncodingType == "url" {
-		err = decodeDeleteObjectsOutput(output)
-		if err != nil {
-			doLog(LEVEL_ERROR, "Failed to get DeleteObjectsOutput with error: %v.", err)
-			output = nil
-		}
-	}
-	return
-}
-
-// SetObjectAcl sets ACL for an object.
-//
-// You can use this API to set the ACL for an object in a specified bucket.
-func (obsClient ObsClient) SetObjectAcl(input *SetObjectAclInput, extensions ...extensionOptions) (output *BaseModel, err error) {
-	if input == nil {
-		return nil, errors.New("SetObjectAclInput is nil")
-	}
-	output = &BaseModel{}
-	err = obsClient.doActionWithBucketAndKey("SetObjectAcl", HTTP_PUT, input.Bucket, input.Key, input, output, extensions)
-	if err != nil {
-		output = nil
-	}
-	return
-}
-
-// GetObjectAcl gets the ACL of an object.
-//
-// You can use this API to obtain the ACL of an object in a specified bucket.
-func (obsClient ObsClient) GetObjectAcl(input *GetObjectAclInput, extensions ...extensionOptions) (output *GetObjectAclOutput, err error) {
-	if input == nil {
-		return nil, errors.New("GetObjectAclInput is nil")
-	}
-	output = &GetObjectAclOutput{}
-	err = obsClient.doActionWithBucketAndKey("GetObjectAcl", HTTP_GET, input.Bucket, input.Key, input, output, extensions)
-	if err != nil {
-		output = nil
-	} else {
-		if versionID, ok := output.ResponseHeaders[HEADER_VERSION_ID]; ok {
-			output.VersionId = versionID[0]
-		}
-	}
-	return
-}
-
-// RestoreObject restores an object.
-func (obsClient ObsClient) RestoreObject(input *RestoreObjectInput, extensions ...extensionOptions) (output *BaseModel, err error) {
-	if input == nil {
-		return nil, errors.New("RestoreObjectInput is nil")
-	}
-	output = &BaseModel{}
-	err = obsClient.doActionWithBucketAndKey("RestoreObject", HTTP_POST, input.Bucket, input.Key, input, output, extensions)
-	if err != nil {
-		output = nil
-	}
-	return
-}
-
-// GetObjectMetadata gets object metadata.
-//
-// You can use this API to send a HEAD request to the object of a specified bucket to obtain its metadata.
-func (obsClient ObsClient) GetObjectMetadata(input *GetObjectMetadataInput, extensions ...extensionOptions) (output *GetObjectMetadataOutput, err error) {
-	if input == nil {
-		return nil, errors.New("GetObjectMetadataInput is nil")
-	}
-	output = &GetObjectMetadataOutput{}
-	err = obsClient.doActionWithBucketAndKey("GetObjectMetadata", HTTP_HEAD, input.Bucket, input.Key, input, output, extensions)
-	if err != nil {
-		output = nil
-	} else {
-		ParseGetObjectMetadataOutput(output)
-	}
-	return
-}
-
-// GetObject downloads object.
-//
-// You can use this API to download an object in a specified bucket.
-func (obsClient ObsClient) GetObject(input *GetObjectInput, extensions ...extensionOptions) (output *GetObjectOutput, err error) {
-	if input == nil {
-		return nil, errors.New("GetObjectInput is nil")
-	}
-	output = &GetObjectOutput{}
-	err = obsClient.doActionWithBucketAndKey("GetObject", HTTP_GET, input.Bucket, input.Key, input, output, extensions)
-	if err != nil {
-		output = nil
-	} else {
-		ParseGetObjectOutput(output)
-	}
-	return
-}
-
-// PutObject uploads an object to the specified bucket.
-func (obsClient ObsClient) PutObject(input *PutObjectInput, extensions ...extensionOptions) (output *PutObjectOutput, err error) {
-	if input == nil {
-		return nil, errors.New("PutObjectInput is nil")
-	}
-
-	if input.ContentType == "" && input.Key != "" {
-		if contentType, ok := mimeTypes[strings.ToLower(input.Key[strings.LastIndex(input.Key, ".")+1:])]; ok {
-			input.ContentType = contentType
-		}
-	}
-	output = &PutObjectOutput{}
-	var repeatable bool
-	if input.Body != nil {
-		_, repeatable = input.Body.(*strings.Reader)
-		if input.ContentLength > 0 {
-			input.Body = &readerWrapper{reader: input.Body, totalCount: input.ContentLength}
-		}
-	}
-	if repeatable {
-		err = obsClient.doActionWithBucketAndKey("PutObject", HTTP_PUT, input.Bucket, input.Key, input, output, extensions)
-	} else {
-		err = obsClient.doActionWithBucketAndKeyUnRepeatable("PutObject", HTTP_PUT, input.Bucket, input.Key, input, output, extensions)
-	}
-	if err != nil {
-		output = nil
-	} else {
-		ParsePutObjectOutput(output)
-	}
-	return
-}
-
-func (obsClient ObsClient) getContentType(input *PutObjectInput, sourceFile string) (contentType string) {
-	if contentType, ok := mimeTypes[strings.ToLower(input.Key[strings.LastIndex(input.Key, ".")+1:])]; ok {
-		return contentType
-	}
-	if contentType, ok := mimeTypes[strings.ToLower(sourceFile[strings.LastIndex(sourceFile, ".")+1:])]; ok {
-		return contentType
-	}
-	return
-}
-
-func (obsClient ObsClient) isGetContentType(input *PutObjectInput) bool {
-	if input.ContentType == "" && input.Key != "" {
-		return true
-	}
-	return false
-}
-
-// PutFile uploads a file to the specified bucket.
-func (obsClient ObsClient) PutFile(input *PutFileInput, extensions ...extensionOptions) (output *PutObjectOutput, err error) {
-	if input == nil {
-		return nil, errors.New("PutFileInput is nil")
-	}
-
-	var body io.Reader
-	sourceFile := strings.TrimSpace(input.SourceFile)
-	if sourceFile != "" {
-		fd, _err := os.Open(sourceFile)
-		if _err != nil {
-			err = _err
-			return nil, err
-		}
-		defer func() {
-			errMsg := fd.Close()
-			if errMsg != nil {
-				doLog(LEVEL_WARN, "Failed to close file with reason: %v", errMsg)
-			}
-		}()
-
-		stat, _err := fd.Stat()
-		if _err != nil {
-			err = _err
-			return nil, err
-		}
-		fileReaderWrapper := &fileReaderWrapper{filePath: sourceFile}
-		fileReaderWrapper.reader = fd
-		if input.ContentLength > 0 {
-			if input.ContentLength > stat.Size() {
-				input.ContentLength = stat.Size()
-			}
-			fileReaderWrapper.totalCount = input.ContentLength
-		} else {
-			fileReaderWrapper.totalCount = stat.Size()
-		}
-		body = fileReaderWrapper
-	}
-
-	_input := &PutObjectInput{}
-	_input.PutObjectBasicInput = input.PutObjectBasicInput
-	_input.Body = body
-
-	if obsClient.isGetContentType(_input) {
-		_input.ContentType = obsClient.getContentType(_input, sourceFile)
-	}
-
-	output = &PutObjectOutput{}
-	err = obsClient.doActionWithBucketAndKey("PutFile", HTTP_PUT, _input.Bucket, _input.Key, _input, output, extensions)
-	if err != nil {
-		output = nil
-	} else {
-		ParsePutObjectOutput(output)
-	}
-	return
-}
-
-// CopyObject creates a copy for an existing object.
-//
-// You can use this API to create a copy for an object in a specified bucket.
-func (obsClient ObsClient) CopyObject(input *CopyObjectInput, extensions ...extensionOptions) (output *CopyObjectOutput, err error) {
-	if input == nil {
-		return nil, errors.New("CopyObjectInput is nil")
-	}
-
-	if strings.TrimSpace(input.CopySourceBucket) == "" {
-		return nil, errors.New("Source bucket is empty")
-	}
-	if strings.TrimSpace(input.CopySourceKey) == "" {
-		return nil, errors.New("Source key is empty")
-	}
-
-	output = &CopyObjectOutput{}
-	err = obsClient.doActionWithBucketAndKey("CopyObject", HTTP_PUT, input.Bucket, input.Key, input, output, extensions)
-	if err != nil {
-		output = nil
-	} else {
-		ParseCopyObjectOutput(output)
-	}
-	return
-}
-
-// AbortMultipartUpload aborts a multipart upload in a specified bucket by using the multipart upload ID.
-func (obsClient ObsClient) AbortMultipartUpload(input *AbortMultipartUploadInput, extensions ...extensionOptions) (output *BaseModel, err error) {
-	if input == nil {
-		return nil, errors.New("AbortMultipartUploadInput is nil")
-	}
-	if input.UploadId == "" {
-		return nil, errors.New("UploadId is empty")
-	}
-	output = &BaseModel{}
-	err = obsClient.doActionWithBucketAndKey("AbortMultipartUpload", HTTP_DELETE, input.Bucket, input.Key, input, output, extensions)
-	if err != nil {
-		output = nil
-	}
-	return
-}
-
-// InitiateMultipartUpload initializes a multipart upload.
-func (obsClient ObsClient) InitiateMultipartUpload(input *InitiateMultipartUploadInput, extensions ...extensionOptions) (output *InitiateMultipartUploadOutput, err error) {
-	if input == nil {
-		return nil, errors.New("InitiateMultipartUploadInput is nil")
-	}
-
-	if input.ContentType == "" && input.Key != "" {
-		if contentType, ok := mimeTypes[strings.ToLower(input.Key[strings.LastIndex(input.Key, ".")+1:])]; ok {
-			input.ContentType = contentType
-		}
-	}
-
-	output = &InitiateMultipartUploadOutput{}
-	err = obsClient.doActionWithBucketAndKey("InitiateMultipartUpload", HTTP_POST, input.Bucket, input.Key, input, output, extensions)
-	if err != nil {
-		output = nil
-	} else {
-		ParseInitiateMultipartUploadOutput(output)
-		if output.EncodingType == "url" {
-			err = decodeInitiateMultipartUploadOutput(output)
-			if err != nil {
-				doLog(LEVEL_ERROR, "Failed to get InitiateMultipartUploadOutput with error: %v.", err)
-				output = nil
-			}
-		}
-	}
-	return
-}
-
-// UploadPart uploads a part to a specified bucket by using a specified multipart upload ID.
-//
-// After a multipart upload is initialized, you can use this API to upload a part to a specified bucket
-// by using the multipart upload ID. Except for the last uploaded part whose size ranges from 0 to 5 GB,
-// sizes of the other parts range from 100 KB to 5 GB. The upload part ID ranges from 1 to 10000.
-func (obsClient ObsClient) UploadPart(_input *UploadPartInput, extensions ...extensionOptions) (output *UploadPartOutput, err error) {
-	if _input == nil {
-		return nil, errors.New("UploadPartInput is nil")
-	}
-
-	if _input.UploadId == "" {
-		return nil, errors.New("UploadId is empty")
-	}
-
-	input := &UploadPartInput{}
-	input.Bucket = _input.Bucket
-	input.Key = _input.Key
-	input.PartNumber = _input.PartNumber
-	input.UploadId = _input.UploadId
-	input.ContentMD5 = _input.ContentMD5
-	input.SourceFile = _input.SourceFile
-	input.Offset = _input.Offset
-	input.PartSize = _input.PartSize
-	input.SseHeader = _input.SseHeader
-	input.Body = _input.Body
-
-	output = &UploadPartOutput{}
-	var repeatable bool
-	if input.Body != nil {
-		_, repeatable = input.Body.(*strings.Reader)
-		if _, ok := input.Body.(*readerWrapper); !ok && input.PartSize > 0 {
-			input.Body = &readerWrapper{reader: input.Body, totalCount: input.PartSize}
-		}
-	} else if sourceFile := strings.TrimSpace(input.SourceFile); sourceFile != "" {
-		fd, _err := os.Open(sourceFile)
-		if _err != nil {
-			err = _err
-			return nil, err
-		}
-		defer func() {
-			errMsg := fd.Close()
-			if errMsg != nil {
-				doLog(LEVEL_WARN, "Failed to close file with reason: %v", errMsg)
-			}
-		}()
-
-		stat, _err := fd.Stat()
-		if _err != nil {
-			err = _err
-			return nil, err
-		}
-		fileSize := stat.Size()
-		fileReaderWrapper := &fileReaderWrapper{filePath: sourceFile}
-		fileReaderWrapper.reader = fd
-
-		if input.Offset < 0 || input.Offset > fileSize {
-			input.Offset = 0
-		}
-
-		if input.PartSize <= 0 || input.PartSize > (fileSize-input.Offset) {
-			input.PartSize = fileSize - input.Offset
-		}
-		fileReaderWrapper.totalCount = input.PartSize
-		if _, err = fd.Seek(input.Offset, io.SeekStart); err != nil {
-			return nil, err
-		}
-		input.Body = fileReaderWrapper
-		repeatable = true
-	}
-	if repeatable {
-		err = obsClient.doActionWithBucketAndKey("UploadPart", HTTP_PUT, input.Bucket, input.Key, input, output, extensions)
-	} else {
-		err = obsClient.doActionWithBucketAndKeyUnRepeatable("UploadPart", HTTP_PUT, input.Bucket, input.Key, input, output, extensions)
-	}
-	if err != nil {
-		output = nil
-	} else {
-		ParseUploadPartOutput(output)
-		output.PartNumber = input.PartNumber
-	}
-	return
-}
-
-// CompleteMultipartUpload combines the uploaded parts in a specified bucket by using the multipart upload ID.
-func (obsClient ObsClient) CompleteMultipartUpload(input *CompleteMultipartUploadInput, extensions ...extensionOptions) (output *CompleteMultipartUploadOutput, err error) {
-	if input == nil {
-		return nil, errors.New("CompleteMultipartUploadInput is nil")
-	}
-
-	if input.UploadId == "" {
-		return nil, errors.New("UploadId is empty")
-	}
-
-	var parts partSlice = input.Parts
-	sort.Sort(parts)
-
-	output = &CompleteMultipartUploadOutput{}
-	err = obsClient.doActionWithBucketAndKey("CompleteMultipartUpload", HTTP_POST, input.Bucket, input.Key, input, output, extensions)
-	if err != nil {
-		output = nil
-	} else {
-		ParseCompleteMultipartUploadOutput(output)
-		if output.EncodingType == "url" {
-			err = decodeCompleteMultipartUploadOutput(output)
-			if err != nil {
-				doLog(LEVEL_ERROR, "Failed to get CompleteMultipartUploadOutput with error: %v.", err)
-				output = nil
-			}
-		}
-	}
-	return
-}
-
-// ListParts lists the uploaded parts in a bucket by using the multipart upload ID.
-func (obsClient ObsClient) ListParts(input *ListPartsInput, extensions ...extensionOptions) (output *ListPartsOutput, err error) {
-	if input == nil {
-		return nil, errors.New("ListPartsInput is nil")
-	}
-	if input.UploadId == "" {
-		return nil, errors.New("UploadId is empty")
-	}
-	output = &ListPartsOutput{}
-	err = obsClient.doActionWithBucketAndKey("ListParts", HTTP_GET, input.Bucket, input.Key, input, output, extensions)
-	if err != nil {
-		output = nil
-	} else if output.EncodingType == "url" {
-		err = decodeListPartsOutput(output)
-		if err != nil {
-			doLog(LEVEL_ERROR, "Failed to get ListPartsOutput with error: %v.", err)
-			output = nil
-		}
-	}
-	return
-}
-
-// CopyPart copy a part to a specified bucket by using a specified multipart upload ID.
-//
-// After a multipart upload is initialized, you can use this API to copy a part to a specified bucket by using the multipart upload ID.
-func (obsClient ObsClient) CopyPart(input *CopyPartInput, extensions ...extensionOptions) (output *CopyPartOutput, err error) {
-	if input == nil {
-		return nil, errors.New("CopyPartInput is nil")
-	}
-	if input.UploadId == "" {
-		return nil, errors.New("UploadId is empty")
-	}
-	if strings.TrimSpace(input.CopySourceBucket) == "" {
-		return nil, errors.New("Source bucket is empty")
-	}
-	if strings.TrimSpace(input.CopySourceKey) == "" {
-		return nil, errors.New("Source key is empty")
-	}
-
-	output = &CopyPartOutput{}
-	err = obsClient.doActionWithBucketAndKey("CopyPart", HTTP_PUT, input.Bucket, input.Key, input, output, extensions)
-	if err != nil {
-		output = nil
-	} else {
-		ParseCopyPartOutput(output)
-		output.PartNumber = input.PartNumber
-	}
-	return
-}
-
 // SetBucketRequestPayment sets requester-pays setting for a bucket.
 func (obsClient ObsClient) SetBucketRequestPayment(input *SetBucketRequestPaymentInput, extensions ...extensionOptions) (output *BaseModel, err error) {
 	if input == nil {
@@ -1276,52 +640,6 @@ func (obsClient ObsClient) GetBucketRequestPayment(bucketName string, extensions
 	if err != nil {
 		output = nil
 	}
-	return
-}
-
-// UploadFile resume uploads.
-//
-// This API is an encapsulated and enhanced version of multipart upload, and aims to eliminate large file
-// upload failures caused by poor network conditions and program breakdowns.
-func (obsClient ObsClient) UploadFile(input *UploadFileInput, extensions ...extensionOptions) (output *CompleteMultipartUploadOutput, err error) {
-	if input.EnableCheckpoint && input.CheckpointFile == "" {
-		input.CheckpointFile = input.UploadFile + ".uploadfile_record"
-	}
-
-	if input.TaskNum <= 0 {
-		input.TaskNum = 1
-	}
-	if input.PartSize < MIN_PART_SIZE {
-		input.PartSize = MIN_PART_SIZE
-	} else if input.PartSize > MAX_PART_SIZE {
-		input.PartSize = MAX_PART_SIZE
-	}
-
-	output, err = obsClient.resumeUpload(input, extensions)
-	return
-}
-
-// DownloadFile resume downloads.
-//
-// This API is an encapsulated and enhanced version of partial download, and aims to eliminate large file
-// download failures caused by poor network conditions and program breakdowns.
-func (obsClient ObsClient) DownloadFile(input *DownloadFileInput, extensions ...extensionOptions) (output *GetObjectMetadataOutput, err error) {
-	if input.DownloadFile == "" {
-		input.DownloadFile = input.Key
-	}
-
-	if input.EnableCheckpoint && input.CheckpointFile == "" {
-		input.CheckpointFile = input.DownloadFile + ".downloadfile_record"
-	}
-
-	if input.TaskNum <= 0 {
-		input.TaskNum = 1
-	}
-	if input.PartSize <= 0 {
-		input.PartSize = DEFAULT_PART_SIZE
-	}
-
-	output, err = obsClient.resumeDownload(input, extensions)
 	return
 }
 
@@ -1408,64 +726,6 @@ func (obsClient ObsClient) GetBucketFetchJob(input *GetBucketFetchJobInput, exte
 	err = obsClient.doActionWithBucketAndKeyV2("GetBucketFetchJob", HTTP_GET, input.Bucket, string(objectKeyAsyncFetchJob)+"/"+input.JobID, input, output, extensions)
 	if err != nil {
 		output = nil
-	}
-	return
-}
-func (obsClient ObsClient) AppendObject(input *AppendObjectInput, extensions ...extensionOptions) (output *AppendObjectOutput, err error) {
-	if input == nil {
-		return nil, errors.New("AppendObjectInput is nil")
-	}
-
-	if input.ContentType == "" && input.Key != "" {
-		if contentType, ok := mimeTypes[strings.ToLower(input.Key[strings.LastIndex(input.Key, ".")+1:])]; ok {
-			input.ContentType = contentType
-		}
-	}
-	output = &AppendObjectOutput{}
-	var repeatable bool
-	if input.Body != nil {
-		_, repeatable = input.Body.(*strings.Reader)
-		if input.ContentLength > 0 {
-			input.Body = &readerWrapper{reader: input.Body, totalCount: input.ContentLength}
-		}
-	}
-	if repeatable {
-		err = obsClient.doActionWithBucketAndKey("AppendObject", HTTP_POST, input.Bucket, input.Key, input, output, extensions)
-	} else {
-		err = obsClient.doActionWithBucketAndKeyUnRepeatable("AppendObject", HTTP_POST, input.Bucket, input.Key, input, output, extensions)
-	}
-	if err != nil {
-		output = nil
-	} else {
-		if err = ParseAppendObjectOutput(output); err != nil {
-			output = nil
-		}
-	}
-	return
-}
-
-func (obsClient ObsClient) ModifyObject(input *ModifyObjectInput, extensions ...extensionOptions) (output *ModifyObjectOutput, err error) {
-	if input == nil {
-		return nil, errors.New("ModifyObjectInput is nil")
-	}
-
-	output = &ModifyObjectOutput{}
-	var repeatable bool
-	if input.Body != nil {
-		_, repeatable = input.Body.(*strings.Reader)
-		if input.ContentLength > 0 {
-			input.Body = &readerWrapper{reader: input.Body, totalCount: input.ContentLength}
-		}
-	}
-	if repeatable {
-		err = obsClient.doActionWithBucketAndKey("ModifyObject", HTTP_PUT, input.Bucket, input.Key, input, output, extensions)
-	} else {
-		err = obsClient.doActionWithBucketAndKeyUnRepeatable("ModifyObject", HTTP_PUT, input.Bucket, input.Key, input, output, extensions)
-	}
-	if err != nil {
-		output = nil
-	} else {
-		ParseModifyObjectOutput(output)
 	}
 	return
 }
