+++ openstack/obs/model_header.go
@@ -1,0 +1,32 @@
+// Copyright 2019 Huawei Technologies Co.,Ltd.
+// Licensed under the Apache License, Version 2.0 (the "License"); you may not use
+// this file except in compliance with the License.  You may obtain a copy of the
+// License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software distributed
+// under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+// CONDITIONS OF ANY KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations under the License.
+package obs
+
+// ISseHeader defines the sse encryption header
+type ISseHeader interface {
+	GetEncryption() string
+	GetKey() string
+}
+
+// SseKmsHeader defines the SseKms header
+type SseKmsHeader struct {
+	Encryption string
+	Key        string
+	isObs      bool
+}
+
+// SseCHeader defines the SseC header
+type SseCHeader struct {
+	Encryption string
+	Key        string
+	KeyMD5     string
+}
